% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapexr-data.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Oncotator annotations for SNVs on chromosome 12 of a PDX sample}
\format{A data frame with 13055 rows and 6 variables
\describe{
  \item{gene}{Gene names (HUGO symbol)}
  \item{chrom}{The chromosome the variant is on}
  \item{loc}{The genomic coordinates of the SNV}
  \item{variant_classification}{The oncotator annotation for the type of mutation (i.e. Missense_Mutation)}
  \item{ref}{The reference allele}
  \item{alt}{The alternate allele}
}}
\usage{
annotations
}
\description{
Oncotator annotations for SNVs on chromosome 12 of a PDX sample
}
\keyword{datasets}
