% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapexr-data.R
\docType{data}
\name{scores}
\alias{scores}
\title{MAPEX scores for the passed variants in \code{variants}}
\format{MAPEX scores
\describe{
  \item{chrom}{The chromosome the variant is on}
  \item{loc}{The genomic coordinates of the SNV}
  \item{variant_score}{The variant score (between 0.0 and 1.0)}
  \item{reason}{The most common classification of reads supporting the variant (mouse,on_target,off_target)}
  }}
\usage{
scores
}
\description{
MAPEX scores for the passed variants in \code{variants}
}
\keyword{datasets}
