% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapexr-data.R
\docType{data}
\name{variants}
\alias{variants}
\title{MuTect1 SNV calls for Chromosome 12 of a PDX sample}
\format{A data frame with 78505 rows and 6 variables
\describe{
  \item{chrom}{The chromosome the variant is on}
  \item{loc}{The genomic coordinates of the SNV}
  \item{ref}{The reference allele}
  \item{alt}{The alternate allele}
  \item{vaf}{Frequency of the alternate allele}
  \item{judgement}{The judgement of MuTect1 on the variant (KEEP/REJECT)}
}}
\usage{
variants
}
\description{
MuTect1 SNV calls for Chromosome 12 of a PDX sample
}
\keyword{datasets}
